Various improvements to code quality for primitives
These are only barely used, so there is limited point to this, but if we
are going to use them they need to be performant. Moving and
re-allocating arrays when unnecessary has a real performance cost, we should avoid
using lists if the size of the collection is known ahead of time.

There was also a decent amount of unused code which we can easily
reimplement if needed, and some methods which were a bit messy for silly
reasons.

There were two PrimitivePacket classes, which worked in very different
ways. I kept both, but made them clearly distinct. (one used indexed
drawing, the other not), and they are both still referenced.