Remove doubled trail logic
Draw is called twice for some reason. There really is no good reason to
do it this way, it is super complicated and unnecessary, from what I can
tell. This looks more or less the same (it might be a little thinner,
because we no longer _draw the same shit twice_, but I think that should
be solved in a different way)

We weren't using the spriteBatch logic in there anyway.

This _will_ fix #34, since the code that caused it no longer exists, but
I cannot guarantee that there won't be some other issue on 64 bit TML.